+++ crates/ruma-client-api/src/error.rs
@@ -147,6 +147,9 @@ pub enum ErrorKind {
     /// M_UNABLE_TO_GRANT_JOIN
     UnableToGrantJoin,
 
+    /// M_BAD_ALIAS
+    BadAlias,
+
     /// FI.MAU.MSC2246_NOT_YET_UPLOADED
     #[cfg(feature = "unstable-msc2246")]
     NotYetUploaded,
@@ -201,6 +204,7 @@ impl AsRef<str> for ErrorKind {
             Self::WeakPassword => "M_WEAK_PASSWORD",
             Self::UnableToAuthorizeJoin => "M_UNABLE_TO_AUTHORISE_JOIN",
             Self::UnableToGrantJoin => "M_UNABLE_TO_GRANT_JOIN",
+            Self::BadAlias => "M_BAD_ALIAS",
             #[cfg(feature = "unstable-msc2246")]
             Self::NotYetUploaded => "FI.MAU.MSC2246_NOT_YET_UPLOADED",
             #[cfg(feature = "unstable-msc2246")]