+++ crates/ruma-client-api/src/lib.rs
@@ -20,6 +20,7 @@ pub mod directory;
 pub mod discovery;
 pub mod error;
 pub mod filter;
+pub mod http_headers;
 pub mod keys;
 pub mod knock;
 pub mod media;