+++ crates/ruma-client-api/CHANGELOG.md
@@ -17,6 +17,7 @@ Improvements:
 * Stabilize support for private read receipts
 * Add support for the pagination direction parameter to `/relations` (MSC3715 / Matrix 1.4)
 * Add support for notifications for threads (MSC3773 / Matrix 1.4)
+* Send CORP headers by default for media responses (MSC3828 / Matrix 1.4)
 
 # 0.15.1